+++ Sources/SystemInfoView.swift
@@ -24,6 +24,8 @@ public struct SystemInfoView: View {
                         .foregroundStyle(currentDevice.thermalState == .nominal ? .primary : currentDevice.thermalState.color, currentDevice.thermalState.color, .secondary)
                     Text(" \(String(describing: currentDevice.thermalState))").font(.headline)
                     Text("thermal state").font(.callout).opacity(0.5)
+                    Text("\(device.cpu.caseName) ")+Text(
+                        Image(symbolName: "cpu"))
                 }
                 Spacer()
                 if let battery = device.battery {